refactor(dashboard): move all components to relevant directories
