+++ apps/dashboard/src/components/ProductActions.vue
