+++ apps/dashboard/src/components/footer/FooterTermsOfServiceModal.vue
