+++ apps/dashboard/src/components/mutations/DepositDetailModal.vue
