+++ apps/dashboard/src/components/mutations/FineDetailModal.vue
