+++ apps/dashboard/src/components/mutations/InvoiceDetailModal.vue
