+++ apps/dashboard/src/components/mutations/TransactionDetailModal.vue
