+++ apps/dashboard/src/components/mutations/WaivedFineDetailModal.vue
