+++ apps/dashboard/src/layout/DashboardLayout.vue
@@ -2,13 +2,13 @@
   <div class="h-screen flex flex-column align-items-center">
     <TopNavbar class="flex-none" />
     <RouterView class="flex-grow-1" />
-    <CopyrightBanner class="flex-none" />
+    <Footer class="flex-none" />
   </div>
 </template>
 
 <script setup lang="ts">
 import { RouterView } from "vue-router";
-import CopyrightBanner from "@/components/CopyrightBanner.vue";
+import Footer from "@/components/footer/Footer.vue";
 import TopNavbar from "@/components/TopNavbar.vue";
 import "@/styles/BasePage.scss";
