+++ apps/dashboard/src/modules/seller/views/ProductsContainersView.vue
@@ -2,7 +2,7 @@
   <div class="page-container">
     <div class="page-title">{{ $t('c_productContainerOperations.Manage Products')}}</div>
     <div class="flex flex-column gap-5">
-      <ContainerCardComponent
+      <ContainerCard
           v-if="containers"
           :header="$t('manageProducts.Containers')"
           :containers="containers"
@@ -15,7 +15,7 @@
 
 <script setup lang="ts">
 import { computed, onBeforeMount } from "vue";
-import ContainerCardComponent from '@/components/ContainerCardComponent.vue';
+import ContainerCard from '@/components/container/ContainerCard.vue';
 import { type ContainerInStore, useContainerStore } from "@/stores/container.store";
 import ProductCard from "@/modules/seller/components/ProductCard.vue";
