+++ apps/dashboard/src/modules/user/views/UserLandingView.vue
@@ -18,7 +18,7 @@
 
 <script setup lang="ts">
 import BalanceWithTopupComponent from '@/modules/user/components/balance/BalanceWithTopup.vue';
-import MutationsTableComponent from '@/components/Mutations/MutationsTableComponent.vue';
+import MutationsTableComponent from '@/components/mutations/MutationsTableComponent.vue';
 import { useAuthStore, useUserStore } from '@sudosos/sudosos-frontend-common';
 import apiService from '@/services/ApiService';
 import type { PaginatedFinancialMutationResponse } from "@sudosos/sudosos-client";
